+++ tools/lgtdoc/xml/lgt2xml.js
@@ -3,6 +3,9 @@
 //   This file is part of Logtalk <http://logtalk.org/>  
 //   Copyright (c) 1998-2014 Paulo Moura <pmoura@logtalk.org>
 // 
+//   XML documenting files to XML conversion script 
+//   Last updated on November 3, 2014
+// 
 //   This program is free software: you can redistribute it and/or modify
 //   it under the terms of the GNU General Public License as published by
 //   the Free Software Foundation, either version 3 of the License, or